Pool gp common to _gp
