--- sva/models/gp/common.py
@@ -1,150 +1,0 @@
-"""Common helper functions for all Gaussian Process models."""
-
-from copy import deepcopy
-
-import gpytorch
-import torch
-from botorch.models import FixedNoiseGP, MultiTaskGP, SingleTaskGP
-from botorch.models.transforms.input import Normalize
-from botorch.models.transforms.outcome import Standardize
-
-
-def set_eval_(gp):
-    """Sets a Gaussian Process model to evaluation mode.
-
-    Parameters
-    ----------
-    gp : botorch.models.SingleTaskGP
-    """
-
-    gp.eval()
-    gp.likelihood.eval()
-
-
-def set_train_(gp):
-    """Sets a Gaussian Process model to training mode.
-
-    Parameters
-    ----------
-    gp : botorch.models.SingleTaskGP
-    """
-
-    gp.train()
-    gp.likelihood.train()
-
-
-def get_model_hyperparameters(model):
-    """Iterates through a torch-like object which has a named_parameters()
-    method and returns a dictionary of all of the parameters, where values
-    are in numpy format.
-
-    Parameters
-    ----------
-    model
-        The model with named_parameters() defined on it.
-
-    Returns
-    -------
-    dict
-    """
-
-    d = {}
-    for p in model.named_parameters():
-        p0 = str(p[0])
-        p1 = p[1].detach().numpy()
-        d[p0] = p1
-    return d
-
-
-def get_simple_model(
-    model_type,
-    X,
-    Y,
-    train_Yvar=None,
-    likelihood=gpytorch.likelihoods.GaussianLikelihood(),
-    mean_module=gpytorch.means.ConstantMean(),
-    covar_module=gpytorch.kernels.ScaleKernel(gpytorch.kernels.MaternKernel()),
-    transform_input=True,
-    transform_output=True,
-    task_feature=-1,
-    **model_kwargs,
-):
-    """A lightweight helper function that returns a botorch SingleTaskGP or
-    FixedNoiseGP."""
-
-    if not isinstance(X, torch.Tensor):
-        X = torch.tensor(X)
-
-    if not isinstance(Y, torch.Tensor):
-        Y = torch.tensor(Y)
-
-    input_transform = (
-        Normalize(X.shape[1], transform_on_eval=True)
-        if transform_input
-        else None
-    )
-    outcome_transform = Standardize(Y.shape[1]) if transform_output else None
-
-    if train_Yvar and not isinstance(train_Yvar, torch.Tensor):
-        train_Yvar = torch.tensor(train_Yvar)
-
-    if model_type == "SingleTaskGP":
-        if train_Yvar is not None:
-            raise ValueError(
-                "You provided train_Yvar but selected SingleTaskGP, which "
-                "does not take this input"
-            )
-
-        model = SingleTaskGP(
-            train_X=X,
-            train_Y=Y,
-            likelihood=likelihood,
-            mean_module=mean_module,
-            covar_module=covar_module,
-            input_transform=input_transform,
-            outcome_transform=outcome_transform,
-            **model_kwargs,
-        )
-
-    elif model_type == "FixedNoiseGP":
-        # Likelihood argument is ignored here
-        model = FixedNoiseGP(
-            train_X=X,
-            train_Y=Y,
-            train_Yvar=train_Yvar,
-            covar_module=covar_module,
-            input_transform=input_transform,
-            outcome_transform=outcome_transform,
-            **model_kwargs,
-        )
-
-    elif model_type == "MultiTaskGP":
-        model = MultiTaskGP(
-            train_X=X,
-            train_Y=Y,
-            train_Yvar=train_Yvar,
-            likelihood=likelihood,
-            mean_module=mean_module,
-            covar_module=covar_module,
-            input_transform=input_transform,
-            outcome_transform=outcome_transform,
-            task_feature=task_feature,
-            **model_kwargs,
-        )
-
-    else:
-        raise ValueError(f"Invalid model type {model_type}")
-
-    return deepcopy(model)
-
-
-class SaveLoadMixin:
-    ...
-    # def save():
-    #     ...
-    #
-    # @classmethod
-    # def load(cls, path):
-    #     d = read_json(path)
-    #     stem = Path(path).stem
-    #     model_path = f"{stem}.pt"
